# api/openapi.yaml
openapi: 3.0.3
info:
  title: Minimal Router OS REST API
  description: Version 1 REST API contract for Minimal Router OS management plane (routerd).
  version: 1.0.0
servers:
  - url: https://192.168.1.1:8443/api/v1
    description: LAN HTTPS management endpoint
  - url: https://10.8.0.1:8443/api/v1
    description: Example authenticated WireGuard management endpoint
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mr_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
paths:
...
  /transactions/pending:
    get:
      summary: Get the provisionally active transaction, if any
      operationId: getPendingTransaction
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Pending transaction metadata without configuration secrets
...